requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
python-dotenv==1.1.1
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==3.2.2
email-validator==2.3.0
sentence-transformers==5.1.0
scikit-learn==1.7.1
PyPDF2==3.0.1
python-docx==1.2.0
opencv-python-headless==4.10.0.84
numpy==2.1.3
faster-whisper==1.0.3
